fix: move PlayerFulfillmentHandler outside of the mixin package
