+++ AGENTS.md
@@ -13,6 +13,7 @@ Uses **Stonecutter** for two versions: `1.21.1-neoforge` (VCS default) and `1.20
 - Build scripts: `build.forge.gradle.kts` / `build.neoforge.gradle.kts`.
 - Custom Gradle plugin `mod-platform` defined in `build-logic/`.
 - When updating adding / removing configuration values in `McTalkingConfig`, make sure to update the `src/main/resources/assets/mc_talking/lang/en_us.json` translation file.
+- ONLY put mixins in the `src/main/java/me/sshcrack/mc_talking/mixin` package. EVERY class in the `mixin` package MUST be a mixin (or accessor).
 
 ## Minecolonies Lookup
 When working with the Minecolonies API, look at the `scripts/MINECOLONIES_DOCS.md` to view the docs and if you need actual code insight, use the gradle classes / minecolonies sources jar to view the source.
